wip: added new fields in the xls and modified the code in every places as required

+++ idas.chox.data/src/main/java/idas/chox/data/ExcelClaim.java
@@ -155,7 +155,7 @@ public class ExcelClaim {
     private String claimVehicleHireVehicleClassName;
     private Date claimVehicleHireRentalStart;
     private Date claimVehicleHireRentalEnd;
-    private Integer claimVehicleOriginalDaysHire;
+    private BigDecimal claimVehicleOriginalDaysHire;
     private Integer claimVehicleHireDays;
     private String claimVehicleHireCollectionReason;
     private String claimVehicleHireHpiVehicleManufacturer;
@@ -426,7 +426,7 @@ public class ExcelClaim {
             claimVehicleHireVehicleClassName = (String) data.get("vh_vehicle_class_name");
             claimVehicleHireRentalStart = (Date) data.get("vh_rental_start");
             claimVehicleHireRentalEnd = (Date) data.get("vh_rental_end");
-            claimVehicleOriginalDaysHire = (Integer) data.get("vh_original_no_days_hire");
+            claimVehicleOriginalDaysHire = (BigDecimal) data.get("vh_original_no_days_hire");
             days = (BigDecimal) data.get("vh_days");
             if (days == null) {
                 claimVehicleHireDays = null;
@@ -1098,7 +1098,7 @@ public class ExcelClaim {
         return claimVehicleHireRentalStart;
     }
 
-    public Integer getClaimVehicleOriginalDaysHire() {
+    public BigDecimal getClaimVehicleOriginalDaysHire() {
         return claimVehicleOriginalDaysHire;
     }
